// src/sentry/static/sentry/app/views/settings/incidentRules/types.tsx
export enum AlertRuleThreshold {
  INCIDENT,
  RESOLUTION,
}

export enum AlertRuleThresholdType {
  ABOVE,
  BELOW,
}

export enum AlertRuleAggregations {
  TOTAL,
  UNIQUE_USERS,
}

export type UnsavedTrigger = {
  alertRuleId: string;
  label: string;
  thresholdType: AlertRuleThresholdType;
  alertThreshold: number;
  resolveThreshold: number;
  timeWindow: number;
};

export type SavedTrigger = UnsavedTrigger & {
  id: string;
  dateAdded: string;
};

export type Trigger = Partial<SavedTrigger> & UnsavedTrigger;

export type UnsavedIncidentRule = {
  aggregations: AlertRuleAggregations[];
  projects: string[];
  query: string;
  timeWindow: number;
  triggers: Trigger[];
};

export type SavedIncidentRule = UnsavedIncidentRule & {
  dateAdded: string;
  dateModified: string;
  id: string;
  status: number;
  name: string;
};

export type IncidentRule = Partial<SavedIncidentRule> & UnsavedIncidentRule;

export enum TimeWindow {
  ONE_MINUTE = 60,
  FIVE_MINUTES = 300,
  TEN_MINUTES = 600,
  FIFTEEN_MINUTES = 900,
  THIRTY_MINUTES = 1800,
  ONE_HOUR = 3600,
  TWO_HOURS = 7200,
  FOUR_HOURS = 14400,
  ONE_DAY = 86400,
}

export type ProjectSelectOption = {
  label: string;
  value: number;
};